test: move realm-operations + published-url tests into realm-server
These pure-logic tests (transport fully mocked) only ran under the
realm-server suite, so the runtime-common SharedTests/runSharedTest
indirection bought nothing. Inline the test bodies as plain qunit modules
in realm-server and drop the runtime-common shared copies.
